updates to Sr2RuO4_SOC
